+++ projects/ide-plugin/Delphi.Terminal.Plugin.DockForm.pas
@@ -15,8 +15,11 @@ unit Delphi.Terminal.Plugin.DockForm;
 interface
 
 uses
-  System.SysUtils, System.Classes,
-  Vcl.Controls, Vcl.Forms, Vcl.ComCtrls,
+  System.SysUtils,
+  System.Classes,
+  Vcl.Controls,
+  Vcl.Forms,
+  Vcl.ComCtrls,
   DockForm,
   Delphi.Terminal.Frame.CmdShell;
 
@@ -41,10 +44,15 @@ type
     function GetCurrentFileDir: string;
     function GetInitialWorkDir: string;
     procedure HandleFormClose(Sender: TObject; var Action: TCloseAction);
+    procedure HandleCommandPaletteRequested(Sender: TObject);
     procedure FocusActiveFrame;
     procedure RegisterIDENotifier;
     procedure UnregisterIDENotifier;
     procedure HandleActiveProjectChanged;
+    function GetActiveProjectFile: string;
+    function GetCurrentFilePath: string;
+    function GetCurrentFileName: string;
+    function ActiveFrame: TframeCmdShell;
   public
     constructor Create(AOwner: TComponent); override;
     destructor Destroy; override;
@@ -58,7 +66,13 @@ type
 implementation
 
 uses
-  Winapi.Windows, ToolsAPI, Delphi.Terminal.Settings;
+  Winapi.Windows,
+  ToolsAPI,
+  Delphi.Terminal.CmdShell,
+  Delphi.Terminal.Settings,
+  Delphi.Terminal.SavedCommands,
+  Delphi.Terminal.VariableExpander,
+  Delphi.Terminal.Plugin.CommandPalette;
 
 type
   TDelphiTerminalIDENotifier = class(TNotifierObject, IOTAIDENotifier)
@@ -85,10 +99,10 @@ begin
     ofnBeginProjectGroupOpen:
       FOwner.FGroupOpening := True;
     ofnEndProjectGroupOpen:
-    begin
-      FOwner.FGroupOpening := False;
-      FOwner.HandleActiveProjectChanged;
-    end;
+      begin
+        FOwner.FGroupOpening := False;
+        FOwner.HandleActiveProjectChanged;
+      end;
     ofnActiveProjectChanged:
       if not FOwner.FGroupOpening then
         FOwner.HandleActiveProjectChanged;
@@ -243,6 +257,7 @@ begin
   AFrame.Align := alClient;
   AFrame.OnRequestProjectDir := HandleRequestProjectDir;
   AFrame.OnRequestFileDir := HandleRequestFileDir;
+  AFrame.OnCommandPaletteRequested := HandleCommandPaletteRequested;
 end;
 
 procedure TfrmDelphiTerminalDock.StartTerminalShell(AFrame: TframeCmdShell; const AShellExe, AWorkDir: string);
@@ -360,6 +375,136 @@ begin
       if FPageControl.ActivePage.Controls[0] is TframeCmdShell then
         TframeCmdShell(FPageControl.ActivePage.Controls[0]).SetWorkingDirectory(ProjectDir);
   end;
+end;
+
+function TfrmDelphiTerminalDock.GetActiveProjectFile: string;
+var
+  ModuleServices: IOTAModuleServices;
+  Project: IOTAProject;
+begin
+  Result := '';
+  if Supports(BorlandIDEServices, IOTAModuleServices, ModuleServices) then
+  begin
+    Project := ModuleServices.GetActiveProject;
+    if Assigned(Project) then
+      Result := Project.FileName;
+  end;
+end;
+
+function TfrmDelphiTerminalDock.GetCurrentFilePath: string;
+var
+  EditorServices: IOTAEditorServices;
+  EditBuffer: IOTAEditBuffer;
+begin
+  Result := '';
+  if Supports(BorlandIDEServices, IOTAEditorServices, EditorServices) then
+  begin
+    EditBuffer := EditorServices.TopBuffer;
+    if Assigned(EditBuffer) then
+      Result := EditBuffer.FileName;
+  end;
+end;
+
+function TfrmDelphiTerminalDock.GetCurrentFileName: string;
+begin
+  Result := ExtractFileName(GetCurrentFilePath);
+end;
+
+function TfrmDelphiTerminalDock.ActiveFrame: TframeCmdShell;
+begin
+  Result := nil;
+  if (FPageControl <> nil) and (FPageControl.ActivePage <> nil) and (FPageControl.ActivePage.ControlCount > 0) then
+    if FPageControl.ActivePage.Controls[0] is TframeCmdShell then
+      Result := TframeCmdShell(FPageControl.ActivePage.Controls[0]);
+end;
+
+function ShellTypeForExe(const AShellExe: string): TSavedCommandShellType;
+var
+  Lower: string;
+begin
+  Lower := LowerCase(ExtractFileName(AShellExe));
+  if Lower.Contains('pwsh') then
+    Result := scPwsh
+  else if Lower.Contains('powershell') then
+    Result := scPowerShell
+  else
+    Result := scCmd;
+end;
+
+procedure TfrmDelphiTerminalDock.HandleCommandPaletteRequested(Sender: TObject);
+var
+  AllCommands, Filtered: TSavedCommandList;
+  ActiveShell: TSavedCommandShellType;
+  ScreenPt: TPoint;
+  PaletteResult: TCommandPaletteResult;
+  Vars: TTerminalVariables;
+  ExpandedCmd, ExpandedDir, CompoundCmd, Unresolved: string;
+  TargetFrame: TframeCmdShell;
+  I: Integer;
+begin
+  TargetFrame := ActiveFrame;
+  if TargetFrame = nil then
+    Exit;
+
+  AllCommands := TerminalSettings.SavedCommands;
+  if AllCommands.Count = 0 then
+    Exit;
+
+  ActiveShell := ShellTypeForExe(TargetFrame.ShellExe);
+  Filtered := TSavedCommandList.Create;
+  try
+    for I := 0 to AllCommands.Count - 1 do
+      if (AllCommands[I].ShellType = scActive) or (AllCommands[I].ShellType = ActiveShell) then
+        Filtered.Add(AllCommands[I]);
+
+    if Filtered.Count = 0 then
+      Exit;
+
+    if Sender is TframeCmdShell then
+      ScreenPt := TframeCmdShell(Sender).ClientToScreen(Point(0, 0))
+    else
+      ScreenPt := ClientToScreen(Point(0, 0));
+
+    PaletteResult := ShowCommandPalette(Self, Filtered, ScreenPt);
+    if PaletteResult.Action = paCancel then
+      Exit;
+  finally
+    Filtered.Free;
+  end;
+
+  Vars := Default(TTerminalVariables);
+  Vars.ProjectDir := GetActiveProjectDir;
+  Vars.ProjectFile := GetActiveProjectFile;
+  Vars.FileDir := GetCurrentFileDir;
+  Vars.FilePath := GetCurrentFilePath;
+  Vars.FileName := GetCurrentFileName;
+  Vars.RadTerminalDir := ExtractFilePath(GetModuleName(HInstance));
+
+  ExpandedCmd := ExpandTerminalVariables(PaletteResult.Command.Command, Vars);
+  ExpandedDir := ExpandTerminalVariables(PaletteResult.Command.WorkingDir, Vars);
+
+  if HasUnresolvedVariables(ExpandedCmd) then
+  begin
+    Unresolved := FindUnresolvedVariable(ExpandedCmd);
+    TargetFrame.ShowMessage(Format('[delphi-terminal] Cannot run "%s": %s could not be resolved', [PaletteResult.Command.Name, Unresolved]));
+    Exit;
+  end;
+  if HasUnresolvedVariables(ExpandedDir) then
+  begin
+    Unresolved := FindUnresolvedVariable(ExpandedDir);
+    TargetFrame.ShowMessage(Format('[delphi-terminal] Cannot run "%s": %s could not be resolved', [PaletteResult.Command.Name, Unresolved]));
+    Exit;
+  end;
+
+  if ExpandedDir <> '' then
+    CompoundCmd := TCmdShellProcess.ChangeDirectoryAndRun(TargetFrame.ShellExe, ExpandedDir, ExpandedCmd)
+  else
+    CompoundCmd := ExpandedCmd;
+
+  if PaletteResult.Action = paEdit then
+    TargetFrame.InsertCommandText(CompoundCmd)
+  else
+    TargetFrame.SendUserCommand(CompoundCmd);
 end;
 
 procedure TfrmDelphiTerminalDock.HandleFormClose(Sender: TObject; var Action: TCloseAction);